+++ pkg/analyzer/CHANGELOG.md
@@ -2,6 +2,7 @@
 * Removed `defaultSdkDirectory()` and `getSdkProperty()` from internal
   `FolderBasedDartSdk`. It is up to the clients to decide how to
   find SDK, for example using `package:cli_util`.
+* Fixed `LintDriver` for the new way to access `WorkspacePackage`.
 
 ## 0.39.13
 * Added 'dart/sdk/build_sdk_summary.dart' with `buildSdkSummary`.